fix(lint): satisfy clang-tidy on device registration + device_id cast
- DeviceMappingManager: static_cast size() to int for rbln_register_device_id
  (bugprone-narrowing-conversions); 3rd arg is `int num_device`
- ProcessGroupRBLN: NOLINT the deliberate signed DeviceIndex->int cast
  (bugprone-signed-char-misuse); the direct cast is intentional per the
  adjacent comment to avoid aliasing negative indices

+++ c10/rbln/DeviceMappingManager.cpp
@@ -159,7 +159,8 @@ void DeviceMappingManager::registerLogicalDevice(int logical_device_index, const
   // Register the logical device with its physical NPU indices
   // Need a non-const copy for rbln_register_device_id which requires int*
   std::vector<int> physical_ids_copy = physical_ids;
-  const int rc = rbln_register_device_id(logical_device_index, physical_ids_copy.data(), physical_ids_copy.size());
+  const int rc = rbln_register_device_id(
+      logical_device_index, physical_ids_copy.data(), static_cast<int>(physical_ids_copy.size()));
   RBLN_CHECK(
       rc == 0,
       "rbln_register_device_id failed for rbln:{} on physical NPU(s) [{}] (rc={}); the device(s) may be in use by "
